+++ main.go
@@ -101,7 +101,7 @@ import (
 )
 
 // ReleaseVersion is the release version for the code.
-var ReleaseVersion = "1.7.0-dev"
+var ReleaseVersion = "1.7.0-rc1"
 
 func main() {
 	exitCode := main2()
@@ -234,7 +234,7 @@ func fetchConfig() error {
 
 	if err := viper.ReadInConfig(); err != nil {
 		switch {
-		case errors.Is(err, viper.ConfigFileNotFoundError{}):
+		case errors.As(err, &viper.ConfigFileNotFoundError{}):
 			// It is allowable for Vouch to not have a configuration file, but only if
 			// we have the information from elsewhere (e.g. environment variables).  Check
 			// to see if we have any beacon nodes configured, as if not we aren't going to
@@ -243,16 +243,10 @@ func fetchConfig() error {
 				// Assume the underlying issue is that the configuration file is missing.
 				return errors.Wrap(err, "could not find the configuration file")
 			}
-		case errors.Is(err, viper.ConfigParseError{}):
+		case errors.As(err, &viper.ConfigParseError{}):
 			return errors.Wrap(err, "could not parse the configuration file")
-		case errors.Is(err, viper.RemoteConfigError("")):
-			return errors.Wrap(err, "could not find the remote configuration file")
-		case errors.Is(err, viper.UnsupportedConfigError("")):
-			return errors.Wrap(err, "unsupported configuration file format")
-		case errors.Is(err, viper.UnsupportedRemoteProviderError("")):
-			return errors.Wrap(err, "unsupported remote configuration provider")
 		default:
-			return err
+			return errors.Wrap(err, "failed to obtain configuration")
 		}
 	}
 
@@ -1211,6 +1205,7 @@ func selectBlindedBeaconBlockProposalProvider(ctx context.Context,
 		blindedBeaconBlockProposalProvider, err = firstblindedbeaconblockproposalstrategy.New(ctx,
 			firstblindedbeaconblockproposalstrategy.WithClientMonitor(monitor.(metrics.ClientMonitor)),
 			firstblindedbeaconblockproposalstrategy.WithLogLevel(util.LogLevel("strategies.blindedbeaconblockproposal.first")),
+			firstblindedbeaconblockproposalstrategy.WithChainTimeService(chainTime),
 			firstblindedbeaconblockproposalstrategy.WithBlindedBeaconBlockProposalProviders(blindedBeaconBlockProposalProviders),
 			firstblindedbeaconblockproposalstrategy.WithTimeout(util.Timeout("strategies.blindedbeaconblockproposal.first")),
 		)