+++ CHANGELOG.md
@@ -10,6 +10,7 @@
   - provide timeout option for remote dirk interactions
   - unblind blocks from all relays that hold the desired payload
   - provide more error information when beacon nodes fail to return expected information
+  - reject blinded blocks with incorrect data before scoring
 
 1.6.3:
   - fix crash attempting to store metrics when prometheus not enabled